+++ enlighten/ui/GUI.py
@@ -5,10 +5,10 @@ from enlighten import common
 
 if common.use_pyside2():
     from PySide2 import QtGui
-    from PySide2.QtWidgets import QMessageBox, QCheckBox
+    from PySide2.QtWidgets import QMessageBox, QCheckBox, QDialog, QLineEdit, QDialogButtonBox, QVBoxLayout, QLabel
 else:
     from PySide6 import QtGui
-    from PySide6.QtWidgets import QMessageBox, QCheckBox
+    from PySide6.QtWidgets import QMessageBox, QCheckBox, QDialog, QLineEdit, QDialogButtonBox, QVBoxLayout, QLabel
 
 log = logging.getLogger(__name__)
 
@@ -149,7 +149,38 @@ class GUI(object):
         dialog.setCheckBox(cb)
         result = dialog.exec_()
 
-        ok = (result == QMessageBox.Ok)
-        checked = dialog.checkBox().isChecked()
+        return { 
+            "ok": result == QMessageBox.Ok,
+            "checked": dialog.checkBox().isChecked()
+        }
 
-        return (ok, checked)
+    def msgbox_with_lineedit_and_checkbox(self, title, label_text, lineedit_text, checkbox_text, checkbox_checked=False):
+        dialog = QDialog(parent=self.ctl.form)
+        dialog.setModal(True)
+        dialog.setWindowTitle(title)
+        dialog.setSizeGripEnabled(True)
+
+        lb = QLabel(label_text, parent=dialog)
+
+        le = QLineEdit(parent=dialog)
+        le.setText(lineedit_text)
+
+        cb = QCheckBox(checkbox_text, parent=dialog)
+        cb.setChecked(checkbox_checked)
+
+        bb = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
+        bb.accepted.connect(dialog.accept)
+        bb.rejected.connect(dialog.reject)
+
+        vb = QVBoxLayout(dialog)
+        vb.addWidget(lb)
+        vb.addWidget(le)
+        vb.addWidget(cb)
+        vb.addWidget(bb)
+
+        result = dialog.exec_()
+        return {
+            "ok": result,
+            "checked": cb.isChecked(),
+            "lineedit": le.text() 
+        }