+++ enlighten/ui/ThumbnailWidget.py
@@ -44,28 +44,17 @@ class ThumbnailWidget(QtWidgets.QFrame):
         log.debug("blocking deep-copy")
 
     def __init__(self, 
-            colors,
-            graph,
-            gui,
-            is_collapsed,
+            ctl,
             measurement,
-            stylesheets,
-            technique,
-            focus_listener,
-            kia = None):
-
+            is_collapsed,
+            graph):
         super(ThumbnailWidget, self).__init__()
 
-        self.colors         = colors
-        self.graph          = graph 
-        self.gui            = gui
-        self.is_collapsed   = is_collapsed
-        self.kia            = kia
+        self.ctl            = ctl
         self.measurement    = measurement
-        self.stylesheets    = stylesheets
-        self.technique      = technique
-        self.focus_listener = focus_listener
-                           
+        self.is_collapsed   = is_collapsed
+        self.graph          = graph 
+
         self.is_displayed  = False
         self.selected_color = None
         self.curve = None
@@ -131,7 +120,7 @@ class ThumbnailWidget(QtWidgets.QFrame):
 
         self.body = QtWidgets.QLabel("loading")
         # self.body.setPixmap(":/simulation/images/spectrums/thumbnail_loading.png")
-        self.stylesheets.apply(self.body, "clear_border")
+        self.ctl.stylesheets.apply(self.body, "clear_border")
         self.body.move(12, 33)
         self.body.resize(162, 112)
         self.body.setScaledContents(True)
@@ -178,7 +167,7 @@ class ThumbnailWidget(QtWidgets.QFrame):
             callback_no     = self.confirm_no_callback,
             callback_yes    = self.confirm_yes_callback,
             parent          = self,
-            stylesheets     = self.stylesheets)
+            stylesheets     = self.ctl.stylesheets)
         self.confirm_widget.move(17, 30)
 
         ########################################################################
@@ -203,16 +192,12 @@ class ThumbnailWidget(QtWidgets.QFrame):
     # ##########################################################################
 
     def should_add_id_button(self):
-        if not self.kia:
-            log.debug("missing kia")
-            return False
-
-        if not self.kia.is_installed:
-            log.debug("kia not installed")
+        if not self.ctl.kia_feature.is_installed:
+            log.debug("KIA not installed")
             return False
 
         # only add "identify" button to Raman measurements
-        return self.technique and "raman" == self.technique.lower()
+        return self.measurement.technique and "raman" == self.measurement.technique.lower()
 
     ##
     # Called by MeasurementFactory to set the rendered thumbnail image
@@ -230,7 +215,7 @@ class ThumbnailWidget(QtWidgets.QFrame):
         policy.setHorizontalPolicy(QtWidgets.QSizePolicy.MinimumExpanding)
 
         le_name = QtWidgets.QLineEdit(self.measurement.label)
-        self.stylesheets.apply(le_name, "clear_border")
+        self.ctl.stylesheets.apply(le_name, "clear_border")
         le_name.setSizePolicy(policy)
         le_name.setMinimumWidth(ThumbnailWidget.MIN_WIDTH)
         le_name.setMaximumWidth(ThumbnailWidget.MAX_WIDTH)
@@ -255,7 +240,7 @@ class ThumbnailWidget(QtWidgets.QFrame):
 
     def set_active(self, flag):
         self.is_displayed = flag
-        self.gui.colorize_button(self.button_display, flag)
+        self.ctl.gui.colorize_button(self.button_display, flag)
 
     def collapse(self):
         self.is_collapsed = True
@@ -320,7 +305,7 @@ class ThumbnailWidget(QtWidgets.QFrame):
         if tooltip is not None:
             button.setToolTip(tooltip)
 
-        self.gui.colorize_button(button, False)
+        self.ctl.gui.colorize_button(button, False)
 
         return button
 
@@ -336,9 +321,8 @@ class ThumbnailWidget(QtWidgets.QFrame):
 
     # the user clicked the "thumbnail" icon, starting an ID operation
     def id_callback(self):
-        if self.kia is not None: 
-            self.kia.enqueue_measurement(self.measurement)
-            self.gui.colorize_button(self.button_id, True)
+        self.ctl.kia_feature.enqueue_measurement(self.measurement)
+        self.ctl.gui.colorize_button(self.button_id, True)
 
     # KIA finished the ID.  For lack of a clear reason to do otherwise, we have
     # the callback come back to the ThumbnailWidget object which originated the 
@@ -352,18 +336,18 @@ class ThumbnailWidget(QtWidgets.QFrame):
         dm = self.measurement.declared_match
         if dm is not None:
             self.button_id.setToolTip("%s (score %.2f)" % (dm, dm.score))
-        self.gui.colorize_button(self.button_id, False)
+        self.ctl.gui.colorize_button(self.button_id, False)
 
     ## Display the ConfirmWidget
     def trash_callback(self):
         self.confirm_widget.setVisible(True)
-        self.gui.colorize_button(self.button_trash, True)
+        self.ctl.gui.colorize_button(self.button_trash, True)
 
     ## 
     # Passed from ConfirmWidget.button_no
     def confirm_no_callback(self):
         self.confirm_widget.setVisible(False)
-        self.gui.colorize_button(self.button_trash, False)
+        self.ctl.gui.colorize_button(self.button_trash, False)
         self.measurement.delete(from_disk=False, update_parent=True)
 
     ##
@@ -393,12 +377,12 @@ class ThumbnailWidget(QtWidgets.QFrame):
         log.debug("rename: now editing")
 
         self.le_name.setReadOnly(False)
-        self.stylesheets.apply(self.le_name, "edit_text")
+        self.ctl.stylesheets.apply(self.le_name, "edit_text")
         self.le_name.setFocus()
         self.le_name.selectAll()
-        self.gui.colorize_button(self.button_edit, True)
+        self.ctl.gui.colorize_button(self.button_edit, True)
 
-        self.focus_listener.register(self.le_name, self.rename_complete_callback)
+        self.ctl.focus_listener.register(self.le_name, self.rename_complete_callback)
 
     def rename_complete_callback(self):
         """
@@ -409,15 +393,12 @@ class ThumbnailWidget(QtWidgets.QFrame):
         log.debug("rename_complete: completing")
         self.last_editted = datetime.now()
 
-        # if not self.focus_listener.registered(self.le_name):
-        #     return
+        self.ctl.focus_listener.unregister(self.le_name)
 
-        self.focus_listener.unregister(self.le_name)
-
-        self.stylesheets.apply(self.le_name, "clear_border")
+        self.ctl.stylesheets.apply(self.le_name, "clear_border")
         self.button_edit.setFocus()
         self.le_name.setReadOnly(True)
-        self.gui.colorize_button(self.button_edit, False)
+        self.ctl.gui.colorize_button(self.button_edit, False)
 
         if self.le_name.text() != self.old_name:
             self.measurement.update_label(self.le_name.text(), manual=True)
@@ -448,7 +429,7 @@ class ThumbnailWidget(QtWidgets.QFrame):
 
     def update_kia(self):
         if self.button_id is not None:
-            self.button_id.setEnabled(self.kia and self.kia.is_installed)
+            self.button_id.setEnabled(self.ctl.kia_feature and self.ctl.kia_feature.is_installed)
 
     ## Retain the thumbnail in the session, but remove its graphical trace from 
     # the chart.  This returns a bool so the clickable "toggle trace" button
@@ -506,16 +487,16 @@ class ThumbnailWidget(QtWidgets.QFrame):
         spectrum = self.measurement.processed_reading.processed
         if self.measurement.processed_reading.is_cropped():
             roi = self.measurement.settings.eeprom.get_horizontal_roi()
-            if roi is not None and self.graph is not None and self.graph.horiz_roi is not None:
+            if roi is not None and self.ctl.horiz_roi is not None:
                 spectrum = self.measurement.processed_reading.processed_cropped
-                x_axis = self.graph.horiz_roi.crop(x_axis, roi=roi)
+                x_axis = self.ctl.horiz_roi.crop(x_axis, roi=roi)
 
         # use named color if found in label
         color = self.selected_color
         if color is None:
-            color = self.colors.color_names.search(label)
+            color = self.ctl.colors.color_names.search(label)
         if color is None:
-            color = self.colors.get_next_random()
+            color = self.ctl.colors.get_next_random()
         pen = pyqtgraph.mkPen(width=1, color=color)
 
         log.debug("add_curve_to_graph: adding name %s, %d y elements, %d x elements (%.2f to %.2f)",
@@ -555,4 +536,4 @@ class ThumbnailWidget(QtWidgets.QFrame):
                     tt += f"{k}: {s}\n"
         # self.body.setToolTip(tt.strip())
         self.body.setWhatsThis(tt.strip())
-        self.stylesheets.apply(self.body, "tooltip")
+        self.ctl.stylesheets.apply(self.body, "tooltip")